add init implementation uratpoly, add mp_pow_ui for rational_class
