+++ symengine/polys/uratpoly.cpp
@@ -1,0 +1,27 @@
+#include <symengine/polys/uratpoly.h>
+
+namespace SymEngine
+{
+
+URatPoly::URatPoly(const RCP<const Basic> &var, URatDict &&dict)
+    : USymEnginePoly(var, std::move(dict))
+{
+    SYMENGINE_ASSERT(is_canonical(poly_))
+}
+
+std::size_t URatPoly::__hash__() const
+{
+    std::size_t seed = URATPOLY;
+
+    seed += var_->hash();
+    for (const auto &it : poly_.dict_) {
+        std::size_t temp = URATPOLY;
+        hash_combine<unsigned int>(temp, it.first);
+        hash_combine<long long int>(temp, mp_get_si(it.second.get_num()));
+        hash_combine<long long int>(temp, mp_get_si(it.second.get_den()));
+        seed += temp;
+    }
+    return seed;
+}
+
+} // SymEngine
