+++ symengine/type_codes.inc
@@ -31,6 +31,7 @@ SYMENGINE_ENUM(MUL, Mul)
 SYMENGINE_ENUM(ADD, Add)
 SYMENGINE_ENUM(POW, Pow)
 SYMENGINE_ENUM(UINTPOLY, UIntPoly)
+SYMENGINE_ENUM(URATPOLY, URatPoly)
 SYMENGINE_ENUM(UEXPRPOLY, UExprPoly)
 SYMENGINE_ENUM(GALOISFIELD, GaloisField)
 SYMENGINE_ENUM(MULTIVARIATEINTPOLYNOMIAL, MultivariateIntPolynomial)
