+++ symengine/visitor.h
@@ -10,6 +10,7 @@
 #include <symengine/polys/uintpoly_piranha.h>
 #include <symengine/polys/uexprpoly.h>
 #include <symengine/polynomial_multivariate.h>
+#include <symengine/polys/uratpoly.h>
 #include <symengine/complex_mpc.h>
 #include <symengine/series_generic.h>
 #include <symengine/series_piranha.h>
